backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.5.3
cloudinary==1.36.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
